bot/requirements.txt
python-telegram-bot==21.10
python-dotenv==1.0.1
fastapi==0.115.6
pandas==2.2.3
numpy==2.2.1
httpx==0.28.1
pydantic-settings==2.7.1
